fix: move toProblemList function to separate utility file and update imports
