+++ src/composables/useProblemSelection.ts
@@ -1,65 +1,9 @@
 import { computed, ref, onMounted } from "vue";
 import api from "@/api";
+import { toProblemList } from "@/utils/normalizeProblemList";
 
 type ProblemSelections = { value: string; text: string }[];
 export type ProblemId2Meta = Record<string, { name: string; quota: number }>;
-
-type ProblemListLike = {
-  count?: number;
-  next?: string | null;
-  previous?: string | null;
-  results?: ProblemItem[];
-  items?: ProblemItem[];
-  data?: {
-    count?: number;
-    next?: string | null;
-    previous?: string | null;
-    results?: ProblemItem[];
-    items?: ProblemItem[];
-  };
-};
-
-function toProblemList(raw: unknown): ProblemList {
-  const data = (raw ?? null) as ProblemListLike | null;
-  let results: ProblemItem[] = [];
-  if (data?.results) {
-    results = data.results;
-  } else if (data?.items) {
-    results = data.items;
-  } else if (data?.data?.results) {
-    results = data.data.results;
-  } else if (data?.data?.items) {
-    results = data.data.items;
-  }
-
-  let count = results.length;
-  if (typeof data?.count === "number") {
-    count = data.count;
-  } else if (typeof data?.data?.count === "number") {
-    count = data.data.count;
-  }
-
-  let next: string | null = null;
-  if (data?.next !== undefined) {
-    next = data.next ?? null;
-  } else if (data?.data?.next !== undefined) {
-    next = data.data.next ?? null;
-  }
-
-  let previous: string | null = null;
-  if (data?.previous !== undefined) {
-    previous = data.previous ?? null;
-  } else if (data?.data?.previous !== undefined) {
-    previous = data.data.previous ?? null;
-  }
-
-  return {
-    count,
-    next,
-    previous,
-    results,
-  };
-}
 
 export function useProblemSelection(courseId: string) {
   const problems = ref<ProblemList>();
